- A proper cube is being rendered (with an improper texture) and rotating right in world space. The cube is loaded from an Object and stored in a global array of all Objects.

+++ Render/D3D11_SurrealRender/D3D11_SurrealRender/D3D11_SurrealRender.cpp
@@ -177,7 +177,7 @@ BOOL InitInstance(HINSTANCE hInstance, int nCmdShow)
    Swap.BufferUsage = DXGI_USAGE_RENDER_TARGET_OUTPUT;
    Swap.SampleDesc.Count = 1;
 
-   MainDisplay.AspectRatio = (Swap.BufferDesc.Width / Swap.BufferDesc.Height);
+   MainDisplay.AspectRatio = ((float)Swap.BufferDesc.Width / (float)Swap.BufferDesc.Height);
 
    UINT flags = D3D11_CREATE_DEVICE_SINGLETHREADED;
 #ifdef _DEBUG
